UPSTREAM: mfd: Add support for the MediaTek MT6359 PMIC
This adds support for the MediaTek MT6359 PMIC. This is a
multifunction device with the following sub modules:

- Codec
- Interrupt
- Regulator
- RTC

It is interfaced to the host controller using SPI interface
by a proprietary hardware called PMIC wrapper or pwrap.
MT6359 MFD is a child device of the pwrap.

Signed-off-by: Hsin-Hsiung Wang <hsin-hsiung.wang@mediatek.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit e545b8f380a96174df40db4203d09156e096ee89)

BUG=b:188596626
TEST=emerge-cherry chromeos-kernel-5_10

Signed-off-by: Fei Shao <fshao@chromium.org>
Change-Id: I5d03aebd39127748247474b5d14001f0a3b70331
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2994307
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>