input: cs40l26: Control to read Power On Sequence
Adds sysfs control power_on_seq that reports the total
number of address-value pairs currently in the
Power On Sequence.

Reading the control also prints the contents of the
Power On Sequence to the dmesg log.

Removes the dev_dbg messages related to the Power On
Sequence in probe.

This is only supported for silicon revision A1 and later.

Change-Id: I6e5e40d8b1eac3181a232847cffa624a05305a10
Signed-off-by: Fred Treven <fred.treven@cirrus.com>