FROMGIT: ath11k: add support for WCN6855
This patch is to add support for WCN6855. For station mode,
WCN6855 is able to connect to an AP, and ping works well.
For AP mode, hostapd is able to bringup an SAP interface with
WCN6855, a normal station can connect to this AP and
ping works well.

Tested-on: WCN6855 hw2.0 PCI WLAN.HSP.1.1-01720.1-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1
Tested-on: QCA6390 hw2.0 PCI WLAN.HST.1.0.1-01740-QCAHSTSWPLZ_V2_TO_X86-1

Signed-off-by: Baochen Qiang <bqiang@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/20210511162214.29475-7-jouni@codeaurora.org
(cherry picked from commit 0fbf19570099cf1c41e86b3b14a392d46131ed0d
 git://git.kernel.org/pub/scm/linux/kernel/git/kvalo/ath.git master)

BUG=b:189366800
TEST=build passes

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: Id4c0e1347ae5d57378f328bd868ee51ebfb43c9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2976048
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>