UPSTREAM: drm/i915/uc: Use platform specific defaults for GuC/HuC enabling
The meaning of 'default' for the enable_guc module parameter has been
updated to accurately reflect what is supported on current platforms.
So start using the defaults instead of forcing everything off.
Although, note that right now, the default is for everything to be off
anyway. So this is not a change for current platforms.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Signed-off-by: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210603164812.19045-2-matthew.brost@intel.com
(cherry picked from commit 47c65b3853f88d105017ef512a521794db51bfeb)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:191366271
TEST=Tested with single/multiple displays, suspend/resume, hotplugs on volteer/zork/trogdor. Build tested on others

Backported using forklift.py

Change-Id: Ife9ed979c220b5b1e6b3c9a53df27996209bf791