UPSTREAM: net: wwan: core: implement TIOCINQ ioctl
It is quite common for a userpace program to fetch the buffered amount
of data in the rx queue to avoid the read block. Implement the TIOCINQ
ioctl to make the migration to the WWAN port usage smooth.

Despite the fact that the read call will return no more data than the
size of a first skb in the queue, TIOCINQ returns the entire amount of
buffered data (sum of all queued skbs). This is done to prevent the
breaking of programs that optimize reading, avoiding it if the buffered
amount of data is too small.

(cherry picked from commit e263c5b2e8912149b49d757511d85a16c5fb432f)

BUG=b:184895604
TEST=Build and boot the kernel with flags CONFIG_WWAN=y and
CONFIG_WWAN_CORE=y and ensure it boots

Signed-off-by: Shrikant Patro <shrikant.patro@intel.com>
Signed-off-by: Dinesh Sharma <dinesh.sharma@intel.com>
Signed-off-by: Sergey Ryazanov <ryazanov.s.a@gmail.com>
Reviewed-by: Loic Poulain <loic.poulain@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Change-Id: Ib30f80767fee4b9bf665ee04be02966e27c15ca6
Signed-off-by: Chiranjeevi Rapolu <chiranjeevi.rapolu@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3047906
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Tested-by: Shrikant Patro <shrikant.patro@intel.corp-partner.google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>