input: cs40l26: Remove silicon rev. checks for algorithm IDs
The CL DSP firmware driver no longer takes the silicon revision
number into account when checking algorithm IDs. This allows for the
CS40L26 driver to avoid checking the firmware mode prior to accessing
firmware controls since the relevant portion of the ID is the same
for all firmware instances.

Change-Id: I7e1382e32baebee5d4d371292ac7448f87029418
Signed-off-by: Fred Treven <fred.treven@cirrus.com>