input: cs40l26: Add control to read die temperature when DSP is active
Adds a sysfs control "die_temp" that reports the current temperature
of the die in 9-bit twos complement form.

The GLOBAL_EN bit must be set for a valid temperature to be reported.
This means that the device must be processing haptic or audio
playback when this control is read.

Change-Id: I29aa2634276e5e0026191337b32787a6234a2ff5
Signed-off-by: Fred Treven <fred.treven@cirrus.com>