input: cs40l26: Implement actuator safe mode workaround
Under specific system-level fault conditions involving a short of the
amplifier output pins to supply, it is possible for the device to short
the supply to ground through the amplifier during
speaker-safe/actuator-safe mode.

This presents a potential risk to the device or the system, and may
result in permanent damage to our device, the system or both.

A single register write needs to be deployed before the amplifier is
enabled to avoid this issue. This register write means the amplifier
outputs will go Hi-Z during the fault condition, and therefore fail-safe.
The consequence of the write is that the amplifier outputs will not be
held at GND during standby and will also be Hi-Z at that time.

The register write is to change SPK_DEFAULT_HIZ to 1, and is achieved by:
Reg 0x7404 = 0x11330000
Note, default for this register is 0x01330000

Change-Id: Ie07cda7c8bc75ac327f97d646584696f756100b1
Signed-off-by: Ben Bright <ben.bright@cirrus.com>