input: cs40l26: Add option to load specified SVC tuning
Use the LE_EST_STATUS firmware control to determine the LE
value of the LRA attached to the output of the device.

If this value falls in a the range specified in the devicetree
for cirrus,svc-le1 or cirrus,svc-le2, load cs40l26-svc1.bin or
cs40l26-svc2.bin respectively. If these properties are omitted
or the LE value is not within the given range, the default
tuning cs40l26-svc.bin will be loaded.

Change-Id: I7461dfade1a6dab2fba79bbd8a771eba2c813e0d
Signed-off-by: Fred Treven <fred.treven@cirrus.com>