BACKPORT: UPSTREAM: Bluetooth: btintel: Add combined setup and shutdown functions
There are multiple setup and shutdown functions for Intel device and the
setup function to be used is depends on the USB PID/VID, which makes
difficult to maintain the code and increases the code size.

This patch adds combined setup and shutdown functions to provide a
single entry point for all Intel devices and choose the setup functions
based on the information read with HCI_Intel_Read_Version command.

Starting from TyP device, the command and response parameters for
HCI_Intel_Read_Version command are changed even though OCF remains
same. However, the legacy devices still can handle the command without
error even if it has a extra parameter, so to simplify the flow,
the new command format is used to read the version information for
both legacy and new (tlv based) format.

Also, it also adds a routine to setup the hdev callbacks in btintel.

Signed-off-by: Tedd Ho-Jeong An <tedd.an@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

(cherry picked from commit ca5425e15881522fed47ddae60c22a487edef182)

Conflicts:
   drivers/bluetooth/btintel.c
   drivers/bluetooth/btintel.h

BUG=b:183594508
TEST=build

Signed-off-by: Michael Sun <michaelfsun@google.com>
Change-Id: Id3e2f5bf6e12a96d8eeb48bac930acc160fd052a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3139414
Reviewed-by: Alain Michaud <alainm@chromium.org>