CHROMIUM: iommu/intel: Specialcase for internal but untrusted devices
Due to security team requirements, we need to treat certain internal
devices (e.g. 5G modem) as untrusted, so that the IOMMU driver uses
bounce buffers for it and uses "strict" setting for it. In future,
this shall be dealt with differently upstream since it will likely
have the feature to switch individual domains to strict or non-strict.

BUG=b:190408519,b:195152713
TEST=Build & check on Brya with 5G SKU that strict settings are used.
Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: Ib770286c9aba5f6839afe7feab82312e9c616399
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3089503
Reviewed-by: Alex Levin <levinale@google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>