input: cs40l26: Use workqueue to start I2S stream
Firmware revision 7.2.7 changes how the DSP handles
the START_I2S mailbox command such that the mailbox
does not clear until the stream has begun.

To avoid a deadlock, send the START_I2S command in
a separate thread by using a workqueue. This allows
the DAPM sequence to complete and for the DSP to
start the incoming stream appropriately.

Change-Id: I5f036f488eebdee3dcd59a03f7eba8a9f551e1b6
Signed-off-by: Fred Treven <fred.treven@cirrus.com>