Revert "CHROMIUM: iommu/intel: Specialcase for internal but untrusted devices"
This reverts commit 588857db26e521674d6759afda9bd54774d9cd41.

This commit was an eyesore, was more intrusive and thus causes conflicts
when we want to cherry-pick IOMMU patches from upstream, and also after
cherry-picking the IOMMU patches, it becomes difficult to apply an
equivalent patch. We shall replace this with another patch that is more
simple & clear.

Signed-off-by: Rajat Jain <rajatja@google.com>

BUG=b:195152713, b:201020414, b:196250702
TEST=build

Change-Id: I0c318e9270fbad3cebfa4c3ffdb4834b6dce0f75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3171208
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Rajat Jain <rajatja@google.com>
Tested-by: Rajat Jain <rajatja@google.com>