input: cs40l26: Delay registration of IRQ until firmware is loaded
The IRQ handler for cs40l26 requires a firmware to be loaded to run.
This patch delays registering the IRQ handler until the firmware has
been loaded, at least in the case of delayed firmware loading. It is
worth noting this fix only covers the delayed case, and probably needs
more work to both improve it and cover the non-delayed path.

Change-Id: I73ce4dcc7a658611c5a21f046df7c1f73532e3ab
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Signed-off-by: Paul Handrigan <Paul.Handrigan@cirrus.com>