input: cs40l26: Add erase and upload operations to ordered workqueue
Adding the driver's functions corresponding to erasing and uploading
effects to the ordered workqueue vibe_workqueue ensures that the
operations will be carried out in the order in which they are placed
on the queue.

This prevents a new upload to commence prior to a previous effect
being erased or stopped, which ocassionally occurred when
delay_before_playback_us was non-zero.

Change-Id: I8d3b5b68332a73be769f4c7d3689426e5f9d2c39
Signed-off-by: Fred Treven <fred.treven@cirrus.com>