input: cs40l26: Replace asp_workqueue with asp_start function
Using a separate workqueue for starting ASP streaming makes it
possible for multiple requests to be made to VIRTUAL1_MBOX_1
which is intended to be solely synchronous.

Replace the workqueue with the function cs40l26_asp_start()
which blocks any further mailbox commands to come in before
START_I2S has been acknowledged by the DSP.

This commit is not functional without the succeeding ASoC
patch that adds this function call to the ASoC driver.

Change-Id: Id974bdb21f1c4b853c83b118bdd49f21812ada3b
Signed-off-by: Fred Treven <fred.treven@cirrus.com>