input: cs40l26: Consolidate PM_TIMER_TIMEOUT code
pm_active_timeout and pm_stdby_timeout are used to set
the time it takes for the DSP to transition from active
to standby and from standby to hibernate respectively.

Familiarize the formats of these two controls
such that a single function is used for writing and
reading to/from the PM_TIMER_TIMEOUT_TICKS firmware
control with the only difference being the register
offset.

Create sysfs controls for reading and changing these
parameters at runtime.

Create devicetree parameters for boottime setting:
"cirrus,pm-active-timeout-ms"
"cirrus,pm-stdby-timeout-ms"
which take the timeout desired for each transition in
milliseconds.

Simplify the code and make clear what each control is
used for. This was somewhat convoluted before.

Change-Id: I5bca7c14b6def01792ef240980e4c0b17e4fe939
Signed-off-by: Fred Treven <fred.treven@cirrus.com>