input: cs40l26: Update error handling procedure
When the device enters Actuator Safe Mode, the DSP
will be guaranteed to be off although the state
variable will not update until ASM is released.

Remove dsp_shutdown() and dsp_wake() in the error
handling procedure, since they are not necessary
even for critical boost errors due to the above
reason.

Change-Id: Iea86f493729ec9f35f5287b15cbb036b1570a70d
Signed-off-by: Fred Treven <fred.treven@cirrus.com>