input: cs40l26: Add DT property "gain-adjust"
Add DT property "cirrus,gain-adjust" which is used in
svc-le* nodes to determine an additional global digital
gain adjustment for ASP streaming if the LE value
matches the range (cirrus,min - cirrus,max) programmed
in the DT.

The value ranges from -100 to 100 and is added to
"cirrus,asp-gain-scale-pct" as the gain to be used
for ASP streaming scenarios. If the adjustment
overshoots the gain limit (100) or is negative the
overall value will be limited to the min/max numbers
automatically.

Change-Id: I38cc7ed754dd136d1ea3e4256fc93d7df3ad0e76
Signed-off-by: Fred Treven <fred.treven@cirrus.com>