input: cs40l26: Move BST_IPK_FLAG unmask to proper function
The update_mask() call for BST_IPK_FLAG must be made any time
firmware is loaded, not just when the input subystem has
initialized.

This bit must be unmasked to avoid sign-extension into the most
significant bits of the IRQ_MASK register when storing the contents
in the power on sequence.

Change-Id: I6c0605a8df3eb6521d2f77dc5af72d97aaa5be72
Signed-off-by: Fred Treven <fred.treven@cirrus.com>