input: cs40l26: Avoid hibernation when initializing PSEQ
If the device hibernates while the Power On Sequence is being
populated, it's possible that it is not configured properly
and data is lost. Set the STANDBY->HIBERNATE timeout to
maximum to avoid inadvertantly hibernating in between
configuration writes. Set the timeout value to the original
value after configuration is complete.

Include the VIRTUAL_2_MBOX_WR interrupt unmask in the same
command as critical error interrupts.

Change-Id: I5f7ab9611a0f88d8e5f6dbf023ac07a008514dc2
Signed-off-by: Fred Treven <fred.treven@cirrus.com>