input: cs40l26: Read IRQ1_STATUS after PM Runtime resume
In rare cases, an interrupt can be sent while a previous
PM runtime call is being made. If the check of IRQ1_STATUS
when entering the interrupt handler is outside of the
PM Runtime framework, the device could be in an unexpected
state with respect to its Power Management.

Avoid this situation by moving the read of IRQ1_STATUS
into the PM runtime framework.

Change-Id: I29ded42d6cd31d337a2e48db7a64abdf0b3c88e0
Signed-off-by: Fred Treven <fred.treven@cirrus.com>