input: cs40l26: Add DT boolean "cirrus,vibe-state"
If the user's firmware version does not include the
DSP to host "CP_TRIGGER" mailbox command, the driver's
reporting of vibe_state is prone to race conditions
and can be set incorrectly.

Users who wish to be able to read from the vibe_state
sysfs control must have "cirrus,vibe-state" present in
their devicetree, and be using the correct firmware.

Any users who do not use this version of firmware must
leave this out of the devicetree for proper operation.

Change-Id: I0224e1171bf5b3c449b88d0581348ba18fe0497d
Signed-off-by: Fred Treven <fred.treven@cirrus.com>