input: cs40l26: Add DT boolean "cirrus,bst-expl-mode-disable"
Add devictree boolean "cirrus,bst-expl-mode-disable" which, if
present, will disable the device's LBST exploratory mode.

This mode allows the boost value to be set greater than the
user's programmed limit. This has the potential to damage the
boost inductor.

To avoid this, the user can use "cirrus,bst-expl-mode-disable",
however, this has the side effect of preventing the LBST short
error interrupt to be detected.

Change-Id: Ia50a188a4bba0decb58d373ad2198c7558969d8f
Signed-off-by: Fred Treven <fred.treven@cirrus.com>