input: cs40l26: Fix value to disable gpi triggers
When an effect is erased, the driver checks if that effect had been
mapped to a gpi event. If so, the driver clears the event table entry
for that gpi event. The correct way to "clear" the event table entry is
to set the register to 0x0001FF as rather that setting to 0x000000.

Change-Id: I3adc6a050d4fddaf763bd8646f0d547404044cf4
Signed-off-by: Ben Bright <ben.bright@cirrus.com>