input: cs40l26: Update PM Transition Timeout Limits
Allow 0 ms to be the minimum timeout value for
ACTIVE->STANDBY and STANDBY->HIBERNATE transitions.
Allow 10000 ms to be the maximum timeout value for
these same transitions.

Set the default timeout value of STANDBY->HIBERNATE transition
to 100 ms and the ACTIVE->STANDBY to 250 ms.

Ensure the maximum value is used when the driver initializes the
power on sequencer.

Change-Id: Ib77b77d054c688e0ace12ba523eadfca9f6e39bd
Signed-off-by: Fred Treven <fred.treven@cirrus.com>