ASoC: cs40l26: Remove check of dai->id in set_tdm_slot()
The set_tdm_slot() callback was return a negative ERRNO
if the snd_soc_dai instance being referenced did not have
an 'id' parameter equal to 0.

This could cause the playback to fail if there are multiple
snd_soc_dai structs registered with the kernel, and the check
does not serve any discernible purpose.

Change-Id: I30bca1bfb281ab891a7642d434d5624ee164bae4
Signed-off-by: Fred Treven <fred.treven@cirrus.com>