input: cs40l26: Update reporting of calibration time
The formula used for calculating the f0_and_q_cal_time_ms
sysfs control was incorrect.

Update the formula and adapt to the firmware's intended usage
which is to use TONE_DURATION_MS to report the calibration
time if this parameter has been populated by the calibration
tuning file (cs40l26-calib.bin).

If TONE_DURATION_MS is 0, use the formula to get the value
from FREQ_SPAN and FREQ_CENTRE.

Change-Id: Ic32d3eb5322faf8a10074b9d8047070d7c31445c
Signed-off-by: Fred Treven <fred.treven@cirrus.com>