ASoC: cs35l43: Add Hibernate DAPM widget
This is a work around for an issue found on Qcom BSP
based on kernel v4.19.
In some use cases, like switching an audio route from
the speaker to the headset and back to the speaker during
a playback in a time smaller than the auto suspend timeout,
the DAPM supply widgets are enabled before the AMP pm runtime
resume is called. So, the audio settings are applied while the AMP
is still hibernated. This change adds a supply widget to exit from
hibernation state if the pm_runtime resume is delayed.

Change-Id: I8110748c6b98b9b27a87a2719861a05e57c114a0
Signed-off-by: Vlad Karpovich <Vlad.Karpovich@cirrus.com>