input: cs40l26: Move stop delay outside mutex
If a stop command is issued immediately after
playing an SVC effect, the SVC initialization
may not have time to complete and thus the
vibe state may not be set before the lock,
resulting in a dropped stop command.

Move the delay_before_stop_playback_us sleep
outside the mutex to allow time for the SVC
initialization phase to complete and the vibe
state to be set.

Change-Id: I0605fdde46e0cf23eed06d060555b301217db86d
Signed-off-by: James Ogletree <james.ogletree@cirrus.com>