input: cs40l26: Avoid DSP lock during driver probe
If a haptic effect is triggered after the driver has
issued a hard reset, there is a possibility that the
DSP becomes stuck in the ACTIVE state.

This prevents the driver from shutting down CCM_CORE_CONTROL
and the initialization process is aborted.

Add a longer delay after the hard reset, and set the
REFCLK to open loop. This will allow the driver
to gracefully recover from the condition.

Disable GPI events from triggering haptic effects; this
will make the scenario less likely to occur.

Change-Id: Iaee9c93dc8ee9f3555832f6dc1a27bd7d3bb6a4f
Signed-off-by: Fred Treven <fred.treven@cirrus.com>