input: cs40l26: Address PSEQ WRITE_FULL issue
There is a bug in the ROM such that when the DSP power-on write
sequencer (aka PSEQ) executes the WRITE_REG_FULL operation, it
sign-extendeds bit 23 into bits [31:24] if bit 23 is set to 1. This
patch replaces instances of WRITE_REG_FULL where this bug is an issue
with a combination of a WRITE_REG_H16 operation and a WRITE_REG_L16
operation. It also issues a warning when a WRITE_REG_FULL operation is
added to the PSEQ where bit 23 of the data is set to 1 and bits 31:24 of
the data are not equal to 0xFF.

Change-Id: Ice3a8e83c80868b034412b7c178ca5f2359e8445
Signed-off-by: Ben Bright <ben.bright@cirrus.com>