input: cs40l26: Ensure effect to be erased is not in-flight
If the driver issues an erase command for an effect
during playback of that same effect, it will erase the
effect internally but the firmware will reject it.
This results in an effective mailbox deadlock as the
driver and firmware wavetables will be out of sync.

Driver should stop the waveform and wait for
stop completion before attempting to erase.

Change-Id: Ifd10de10a5e59a25ad0da1aa4f18c0a34e9aad44
Signed-off-by: James Ogletree <james.ogletree@cirrus.com>