input: cs40l26: Add support for DVL PEQ cal request
DVL PEQ calibration requires that the user provide a calibrated F0
value. This patch includes a change that copies the calibrated F0_EST
value over to the DVL LRA_NORM_F0 control. This allows this requirement
to be transparent to the user. The user simply needs to request F0
calibration and then request DVL PEQ calibration.

Change-Id: I4e5efc90c72a8383ee008ccfe5cf852a9bd0b40b
Signed-off-by: Ben Bright <ben.bright@cirrus.com>