input: cs40l26: Add support for reporting time to DSP for EP
As part of Excursion Protection (EP), the DSP models the temperature of
the LRA and the LRA coil. The DSP does not keep a timer running
while in hibernation, so the driver must communicate the time elapsed
while in hibernation. To do so, the driver uses the new HE_TIME mailbox
command.

HE_TIME Payload:
[23] overflow
when this bit is set, the time elapsed since the sending
of the most recent ALLOW_HIBERNATE command is greater than 0x7FFFFF
(8388607ms = ~140 min)
[22:0] time_elapsed_ms
when overflow is not set, this field contains
the time elapsed in milliseconds since the most recent ALLOW_HIBERNATE
mailbox command was sent

Change-Id: I7eecaedd046be85438bdc0d55d2bdd200306fba3
Signed-off-by: Ben Bright <ben.bright@cirrus.com>