firmware: cirrus: Ensure allocation for NULL terminator.
info_str was not accounting for the NULL terminator when
calculating the size of the necessary memory allocation in
cl_dsp_coeff_handle_info_text and cl_dsp_handle_info_text.

Add 1 to the provided data length to resolve the KASAN warning
associated with this bug and ensure allocation includes space
for the string's NULL terminator.

Change-Id: I8cad53d9d03701663e66ffedbf59957682102c12
Signed-off-by: Fred Treven <fred.treven@cirrus.com>