input: cs40l26: Enable cl_dsp debugfs for deferred F/W case
When cirrus,fw-defer is present in the cs40l26 devicetree node,
the driver waits for the user to manually trigger firmware loading.
This usecase was causing the CL DSP Debugfs setup to fail with no
way to enable it at a later time.

Resolve this issue by moving the debugfs setup code to within
the firmware loading/configuration sequence.

Change-Id: I213f7c427f96326191b747fd0721dbf664db7a20
Signed-off-by: Fred Treven <fred.treven@cirrus.com>