power: supply: ln8411: Applies OTP via regmap_register_patch
Uses regmap_register_patch to apply an OTP table to keep the regmap
cache in sync. Our OTP is unique and can be used to alter the POR
values of user bits. For A1 there is a single OTP table, but for B0
there may be more than one table.

Change-Id: Iab7fc547d14a88787afe51cf1f3b49769c8adf24
Signed-off-by: Ricardo Rivera-Matos <rriveram@opensource.cirrus.com>