input: cs40l26: Set minimum limit on standby to hibernate timeout
The pm_stdby_timeout_ms sysfs control and the
"cirrus,pm-stdby-timeout-ms" devicetree control are used to
configure the firmware's setting for how long it takes the device
to transition from a standby state (part is idle and not actively
processing) to a hibernation state (part is idle and in low-power mode).

Hardware contents are lost when the device enters hibernation.
A race condition is present if the timeout value is too low
because information from the DSP to host mailbox may be lost
before the driver is able to prevent hibernation.

Avoid this by enforcing a minimum value of 100 ms for the standby
to hibernate transition time.

Improve the code for active and standby timeout values by combining the
configuration into a single function.

Change-Id: Ibf9a1bf7392b4d283cf4531d11e776bbb1cec321
Signed-off-by: Fred Treven <fred.treven@cirrus.com>