Input: cs40l26 - Change value to disable boost exploratory mode
The value being written to TEST_LBST in order to disable boost
exploratory mode was set to 0x014DC080 but bits [23:16] are
undefined and only bit 24 is relevant for enable/disable of
this functionality.

Avoid confusion by changing this value to 0x0100C080.

Add test key lock code to A1 errata writes in order to
properly lock the test key after the TEST_LBST configuration
is complete.

Change-Id: I68a5c294d687e31d5662606883c7d146a544d26c
Signed-off-by: Fred Treven <fred.treven@cirrus.com>