dt-bindings: input: cirrus,cs40l26: Update documentation for gpi default mapping
Change format of cirrus,release-index and cirrus,press-index to be an
array of uint32 values rather than one string and one integer.

This enables the driver to avoid using any of_* functions which are
being deprecated and instead allows for device_* functions when
parsing the devicetree.

The RAM waveform bank corresponds to a value of 0, and the ROM
waveform bank corresponds to a value of 1.

Change-Id: Ib7a6a41b2f462cd4e443f0fd34fc7bdfb7eadf69
Signed-off-by: Fred Treven <fred.treven@cirrus.com>