Input: cs40l26 - Remove use of platform_data
Read devicetree properties without using intermediary
platform_data struct.

Removes use of any of_* functions which will soon be
deprecated, and rely instead on device_property_*
or fwnode_* functions.

Improves brownout prevention implementation  in terms
of readability and fixes a bug in which the wrong value
was being used to check if brownout threshold was
out of bounds.

Change-Id: I2beb4983c93567f046b1ef34964dba16d472f420
Signed-off-by: Fred Treven <fred.treven@cirrus.com>