firmware: cs_dsp: Fix new control name check
Before adding a new FW control, its name is checked against
existing controls list. But the string length in strncmp used
to compare controls names is taken from the list, so if beginnings
of the controls are matching,  then the new control is not created.
For example, if CAL_R control already exists, CAL_R_SELECTED
is not created.
The fix is to compare string lengths as well.

Change-Id: I70a48513d1cb17520eb91e3fde4a8bf2a631d401
Signed-off-by: Vlad Karpovich <vkarpovi@opensource.cirrus.com>