Input: cs40l26 - Ensure proper range checking for gain adjust
"cirrus,gain-adjust" from the "svc-le" devicetree node has a
range of -100 to 100. The range checking in the driver when
determining whether or not to program the gain adjust was treating
this as an unsigned value which would reject any negative values.

Change the check to use a signed variable so that the proper limits
can be enforced.

Change-Id: I6d6a47f8e22c29e5f8920d0be9b01bd8af0fbebb
Signed-off-by: Fred Treven <fred.treven@cirrus.com>