dt-bindings: input: cirrus,cs40l26: Remove "multipleOf" restriction
"cirrus,vpbr-thld-uv" and "cirrus,vbbr-thld-uv" values cannot
coincide with the "multipleOf" stipulation from the YAML file
since the accepted values are not multiples of the step sizes.

The values previously used are actually the step sizes for
each control which begin from a non-zero minimum.

Mention the granularity in the description of the controls
rather than using the "multipleOf" parameter.

Change-Id: I6fa65dcdd3309828f5404af397ed598188b8bfaa
Signed-off-by: Fred Treven <fred.treven@cirrus.com>